+++ index.js
@@ -5,12 +5,6 @@ const plugins = require('./src/plugins')
 const rules = require('./src/rules')
 const settings = require('./src/settings')
 
-/**
- * This file returns the configuration for all the eslint plugins.
- *
- * TODO (2021-02-01) Wrap this up in something more than just an Eslint config file, maybe?
- */
-
 module.exports = {
   env: environment,
   extends: [],