+++ src/definition/you-dont-need-momentjs.js
@@ -1,6 +1,0 @@
-module.exports = {
-  'you-dont-need-momentjs/no-dynamic-import-moment': 'error',
-  'you-dont-need-momentjs/no-import-moment': 'error',
-  'you-dont-need-momentjs/no-moment-constructor': 'error',
-  'you-dont-need-momentjs/no-require-moment': 'error'
-}