+++ src/plugins.js
@@ -11,10 +11,6 @@ const plugins = [
   'unicorn',
 ]
 
-if (isInstalled.sync('jest')) plugins.push('jest')
-
 if (isInstalled.sync('lodash')) plugins.push('you-dont-need-lodash-underscore')
-
-if (isInstalled.sync('momentjs')) plugins.push('you-dont-need-momentjs')
 
 module.exports = plugins