+++ src/rules.js
@@ -5,16 +5,8 @@ const isInstalled = require('is-installed')
 const ruleCollection = importModules(resolve(__dirname, './definition/'), { camelize: false })
 const rules = {}
 
-if (!isInstalled.sync('jest')) {
-  delete ruleCollection.jest
-}
-
 if (!isInstalled.sync('lodash')) {
   delete ruleCollection['you-dont-need-lodash-underscore']
-}
-
-if (!isInstalled.sync('momentjs')) {
-  delete ruleCollection['you-dont-need-momentjs']
 }
 
 Object.keys(ruleCollection)